move styles to file; describe function
