+++ src/app/views/query-runner/query-input/auto-complete/suffix/HintList.tsx
@@ -1,14 +1,15 @@
 import { Link, Separator, Text } from '@fluentui/react';
 import React from 'react';
 
-import { IHint, styles } from './suffix-util';
+import { IHint } from './suffix-util';
+import { styles } from './suffix.styles';
 
 export const HintList = ({ hints }: any) => {
   const listItems = hints.map((hint: IHint, index: any) => <div key={index}>
-    {hint.description && <Text block variant="small" id={'description' + index}>
+    {hint.description && <Text block variant='small' id={'description' + index}>
       {hint.description}
     </Text>}
-    {hint.link && <Link href={hint.link.url} target="_blank" className={styles.link}>
+    {hint.link && <Link href={hint.link.url} target='_blank' className={styles.link}>
       {hint.link.name}
     </Link>}
     <Separator />
